fixing json and numpy int issue

+++ converters/detection2panoptic_coco_format.py
@@ -43,8 +43,8 @@ def convert_detection_to_panoptic_coco_format_single_core(
                                                                  working_idx,
                                                                  len(img_ids)))
         img = coco_detection.loadImgs(int(img_id))[0]
-        pan_format = np.zeros((img['height'], img['width'], 3), dtype=np.uint8)
-        overlaps_map = np.zeros((img['height'], img['width']), dtype=np.uint32)
+        pan_format = np.zeros((img['height'], img['width'], 3))
+        overlaps_map = np.zeros((img['height'], img['width']))
 
         anns_ids = coco_detection.getAnnIds(img_id)
         anns = coco_detection.loadAnns(anns_ids)
@@ -90,7 +90,7 @@ def convert_detection_to_panoptic_coco_format_single_core(
         ann['category_id'] = 2
         segment_id, color = id_generator.get_id_and_color(2)
         non_archaeo_mask = (1 - overlaps_map)
-        fortran_ground_truth_binary_mask = np.asfortranarray(non_archaeo_mask, dtype=np.uint8)
+        fortran_ground_truth_binary_mask = np.asfortranarray(non_archaeo_mask, dtype=int)
         encoded_ground_truth = COCOmask.encode(fortran_ground_truth_binary_mask)
         ground_truth_area = COCOmask.area(encoded_ground_truth)
         ground_truth_bounding_box = COCOmask.toBbox(encoded_ground_truth)
